+++ kitty/window_list.py
@@ -513,9 +513,11 @@ class WindowList:
         return {gr.id: ((gr is ag and draw_active_borders) or gr.needs_attention) for gr in self.groups}
 
     @property
-    def num_visble_groups(self) -> int:
+    def has_more_than_one_visible_group(self) -> bool:
         ans = 0
         for gr in self.groups:
             if gr.is_visible_in_layout:
                 ans += 1
-        return ans
+                if ans > 1:
+                    return True
+        return False